backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-multipart>=0.0.6
pandas>=2.1.3
numpy>=1.24.3
scikit-learn>=1.3.2
xgboost>=2.0.2
torch>=2.1.0
transformers>=4.35.0
sentence-transformers>=2.2.2
accelerate>=0.26.0
faiss-cpu>=1.7.4
datasets>=2.14.7
chromadb>=0.4.18
openpyxl>=3.1.2
python-dotenv>=1.0.0
pydantic-extra-types>=2.3.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-postgres>=0.0.3
langchain-openai>=0.0.5
langchain-ollama>=0.1.0
langchain-huggingface>=0.0.1
pgvector>=0.2.3
PyPDF2>=3.0.0
pdfplumber>=0.9.0
Pillow>=10.0.0
pytesseract>=0.3.10

